+++ sdk/src/relay/api/mod.rs
@@ -1,5 +1,6 @@
 mod fetch_events;
 mod send_event;
+mod send_msg;
 mod stream_events;
 mod subscribe;
 mod sync;
@@ -9,6 +10,7 @@ mod unsubscribe_all;
 
 pub use self::fetch_events::*;
 pub use self::send_event::*;
+pub use self::send_msg::*;
 pub use self::stream_events::*;
 pub use self::subscribe::*;
 pub use self::sync::*;